+++ R/deprecated.R
@@ -18,3 +18,31 @@ is.natomic <- function(x) {
   )  
   is_natomic(x)
 }
+
+#' @describeIn deprecated Is nlist
+#'
+#' \lifecycle{soft-deprecated}
+#'
+#' Replace by [is_nlist()]
+#' @export
+is.nlist <- function(x) {
+  deprecate_soft("0.0.1",
+    what = "is.nlist()",
+    with = "is_nlist()"
+  )  
+  is_nlist(x)
+}
+
+#' @describeIn deprecated Is nlists
+#'
+#' \lifecycle{soft-deprecated}
+#'
+#' Replace by [is_nlists()]
+#' @export
+is.nlists <- function(x) {
+  deprecate_soft("0.0.1",
+    what = "is.nlists()",
+    with = "is_nlists()"
+  )  
+  is_nlists(x)
+}